backend/requirements.txt
alembic==1.10.3
fastapi==0.95.1
uvicorn==0.21.1
sqlalchemy==2.0.9
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.6
celery==5.2.7
redis==4.5.4
httpx==0.24.0
pandas==2.0.0
openpyxl==3.1.2
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
openai==0.27.4
gunicorn==20.1.0
pytest==7.3.1
alembic==1.10.3
